+++ libs/core/lib/command-handler.ts
@@ -1,5 +1,0 @@
-import { Command } from './command';
-
-export interface CommandHandler<T extends Command = Command> {
-  handle(command: T): Promise<void>;
-}